requirements.txt
hidapi==0.14.0.post4
setuptools==75.8.2
uv==0.6.3
